+++ packages/solid-router/src/Match.tsx
@@ -460,7 +460,16 @@ export const MatchInner = (): any => {
                   invariant()
                 }
 
-                return null
+                const matchId = currentMatch().id
+                const routerMatch = Solid.untrack(() =>
+                  router.getMatch(matchId),
+                )
+                const loaderResult = Solid.createMemo(async () => {
+                  await new Promise((r) => setTimeout(r, 0))
+                  return getLoadPromise(matchId, routerMatch)
+                })
+
+                return <>{loaderResult()}</>
               }}
             </Solid.Match>
             <Solid.Match when={currentMatch().status === 'error'}>